requirements-monitor.txt
# NECPGAME Performance Monitor Requirements
# Issue: Performance monitoring system dependencies

psutil==5.9.6
requests==2.31.0
pyyaml==6.0.1
psycopg2-binary==2.9.7
python-dateutil==2.8.2
